Break Program component out of ProjectorMain's file
